laba3_5/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
djoser
python-dotenv
psycopg2-binary